+++ src/VM/State.cpp
@@ -2,6 +2,9 @@
 
 #include <Ark/Constants.hpp>
 
+#define _CRT_SECURE_NO_WARNINGS  // avoid warning on getenv
+#include <stdlib.h>
+
 namespace Ark
 {
     State::State(const std::string& libdir, uint16_t options) :
@@ -10,7 +13,11 @@ namespace Ark
     {
         // read environment variable to locate ark std lib
         if (m_libdir == "?")
-        {}
+        {
+            char* val = getenv("ARKSCRIPT_PATH");
+            m_libdir = val == nullptr ? "" : std::string(val);
+            m_libdir += "/lib";
+        }
     }
 
     bool State::feed(const std::string& bytecode_filename)