+++ CHANGELOG.md
@@ -6,7 +6,6 @@
 - `(fill qu value)` create a list of `qu` `value`s
 - `(setListAt list at new-value)` modify a list in place and return the new list value
 - adding UTF-8 support in programs (experimental)
-- adding EASTL as a submodule
 
 ### Changed
 - UserType does not need to be given a manually defined type id but relies on `typeid(T)`
@@ -15,8 +14,10 @@
 - renaming CLI feature options:
     - `-ffunction-arity-check` becomes `-ffac`, same for the `-fno-` version
     - `-fauthorize-invalid-token-after-paren` becomes `-faitap`, some for the `-fno-` version
+- improving compiler performances by using const ref when passing nodes around
 
 ### Removed
+- removed NFT from the internal API to rely only on the value type
 
 ## 3.0.10
 ### Added